# specs/012-server-ingest-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GRAF Server Ingest API
  description: Backend ingest foundation for finalized local recording artifacts.
  version: 0.1.0
paths:
...
  /api/v1/admin/overview:
    get:
      tags:
      - admin
      summary: Get Admin Overview
      operationId: getAdminOverview
      parameters:
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Getadminoverview
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users:
    get:
      tags:
      - admin
      summary: List Admin Users
      operationId: listAdminUsers
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Search
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(owner|admin|member)$
          - type: 'null'
          title: Role
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Status
      - name: invitation_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Invitation Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Listadminusers
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users/{user_id}:
    get:
      tags:
      - admin
      summary: Get Admin User
      operationId: getAdminUser
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Getadminuser
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/invitations:
    post:
      tags:
      - admin
      summary: Create Admin Invitation
      operationId: createAdminInvitation
      parameters:
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Createadmininvitation
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/invitations/{invitation_id}/revoke:
    post:
      tags:
      - admin
      summary: Revoke Admin Invitation
      operationId: revokeAdminInvitation
      parameters:
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Invitation Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationRevokeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Revokeadmininvitation
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/invitations/{invitation_id}/complete:
    post:
      tags:
      - admin
      summary: Complete Admin Invitation
      operationId: completeAdminInvitation
      parameters:
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Invitation Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCompleteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Completeadmininvitation
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InvitationCompleteRequest:
      properties:
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        login_contact:
          type: string
          maxLength: 240
          minLength: 1
          title: Login Contact
        provider:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Provider
      additionalProperties: false
      type: object
      required:
      - workspace_id
      - login_contact
      title: InvitationCompleteRequest
    InvitationCreateRequest:
      properties:
        target_contact:
          type: string
          maxLength: 240
          minLength: 1
          title: Target Contact
        target_provider:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Target Provider
        invited_role:
          type: string
          pattern: ^(owner|admin|member)$
          title: Invited Role
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
      additionalProperties: false
      type: object
      required:
      - target_contact
      - invited_role
      title: InvitationCreateRequest
    InvitationRevokeRequest:
      properties:
        reason_code:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Reason Code
      additionalProperties: false
      type: object
      title: InvitationRevokeRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...